backend/requirements.txt
neo4j
torch 
python-dotenv
transformers
peft
bitsandbytes
protobuf
sentencepiece
flask
flask_cors